removed soil plant hydrology example
