--- experiments/integrated/ozark/hydrology_only/ozark_simulation.jl
@@ -1,10 +1,0 @@
-t0 = FT(140 * 3600 * 24)
-N_days = 140
-tf = t0 + FT(3600 * 24 * N_days)
-dt = FT(225)
-n = 16
-saveat = Array(t0:(n * dt):tf)
-timestepper = CTS.ARS222()
-norm_condition = CTS.MaximumError(FT(1e-8))
-conv_checker = CTS.ConvergenceChecker(; norm_condition)
-max_iterations = 20
